+++ RELEASENOTES.md
@@ -71,6 +71,8 @@
     *   MP4: Improve sniffing efficiency of very large files by assuming a
         `stbl` box larger than 1MB implies the file must be non-fragmented
         ([#2650](https://github.com/androidx/media/issues/2650)).
+    *   Matroska: Add support for DTS-HD detection
+        ([#6225](https://github.com/google/ExoPlayer/issues/6225)).
 *   DataSource:
 *   Audio:
     *   Make `AudioProcessor` instances aware of seeking.